requirements.txt
alembic==1.5.8
apache-airflow==2.0.2
apache-airflow-providers-amazon==1.3.0
apache-airflow-providers-apache-spark==1.0.2
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-postgres==1.0.1
apache-airflow-providers-sftp==1.1.1
apache-airflow-providers-sqlite==1.0.2
apispec==3.3.2
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.3.0
blinker==1.4
boto3==1.17.57
botocore==1.20.57
cached-property==1.5.2
cattrs==1.5.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
dill==0.3.3
dnspython==1.16.0
docutils==0.17.1
email-validator==1.1.2
entrypoints==0.3
eventlet==0.30.2
findspark==1.4.2
Flask==1.1.2
Flask-AppBuilder==3.2.2
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gevent==21.1.2
graphviz==0.16
greenlet==1.0.0
gunicorn==19.10.0
httplib2==0.19.1
idna==2.10
importlib-resources==1.5.0
inflection==0.5.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
mistune==0.8.4
natsort==7.1.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
prison==0.1.3
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py4j==0.10.9
pycparser==2.20
pygit2==1.5.0
Pygments==2.8.1
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pysftp==0.2.9
pyspark==3.1.1
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
rich==9.2.0
s3transfer==0.4.2
Send2Trash==1.5.0
setproctitle==1.2.2
six==1.15.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.0
sshtunnel==0.1.5
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.4
watchtower==0.7.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
